refactor(java/immutable): 단계별로 코드 구분
