+++ java-basic/src/main/java/com/test/java/javatest/immutable/step2/ImmutableMain.java
@@ -1,8 +1,8 @@
-package com.test.java.javatest.immutable;
+package com.test.java.javatest.immutable.step2;
 
 public class ImmutableMain {
     public static void main(String[] args) {
         Person person = new Person("길동이", 20);
-        person.setAge(21);
+//        person.setAge(21);
     }
 }
